media-tv/kodi: fix ffmpeg-compat detection
The right include dir was passed either way, but cmake use has
its own logic (not using pkg-config) and fails to find it if
non-compat ffmpeg is not also installed.

Note it is possible there are other cmake-using packages broken
in a similar way given they almost all do their own (different)
custom thing. Ultimately they'd have needed to be tested with
both installed and not installed to ensure they both do not use
the wrong one if present, and detect the right one properly --
only did the former so far (might review a few).

Thanks-to: Chris Mayo
Closes: https://bugs.gentoo.org/951504
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>